feat: build desktop investor edit profile form
