+++ src/domain/profile/atoms/desktop-investor-edit-profile-form/desktop-investor-edit-profile-form.styles.ts
@@ -1,6 +1,8 @@
 import styled from '@emotion/styled';
 
-const StyledWrapper = styled.div``;
+const StyledWrapper = styled.form`
+  max-width: 1400px;
+`;
 
 const S = {
   StyledWrapper,
