+++ src/domain/profile/atoms/desktop-investor-edit-profile-form/desktop-investor-edit-profile-form.types.ts
@@ -10,7 +10,7 @@ import {
   ITeamSize,
 } from '@interfaces';
 
-export interface IInvestorEditProfileFormProps {
+export interface IDesktopInvestorEditProfileFormProps {
   focusMarkets: IFocusMarket[];
   industrialSectors: IIndustrialSector[];
   investmentSizes: IInvestmentSize[];
