+++ src/domain/profile/atoms/index.ts
@@ -1,2 +1,2 @@
 export * from './section-wrapper/section-wrapper.layout';
-export * from './investor-edit-profile-form/investor-edit-profile-form.component';
+export * from './desktop-investor-edit-profile-form/desktop-investor-edit-profile-form.component';
