+++ src/domain/profile/molecules/index.ts
@@ -1,2 +1,2 @@
-export * from './edit-profile-form/edit-profile-form.component';
 export * from './subscription-plans/subscription-plans.component';
+export * from './desktop-edit-profile-form/desktop-edit-profile-form.component';
