+++ src/domain/profile/organisms/desktop-profile/desktop-profile.component.tsx
@@ -1,10 +1,10 @@
-import { EditProfileForm, SubscriptionPlans } from '../../molecules';
+import { DesktopEditProfileForm, SubscriptionPlans } from '../../molecules';
 import { IProfileProps } from '../../profile.types';
 import S from './desktop-profile.styles';
 
 export const DesktopProfile = ({ plans, ...restProps }: IProfileProps): JSX.Element => (
   <S.StyledWrapper>
     <SubscriptionPlans plans={plans} />
-    <EditProfileForm {...restProps} />
+    <DesktopEditProfileForm {...restProps} />
   </S.StyledWrapper>
 );
