+++ src/domain/profile/organisms/mobile-profile/mobile-profile.component.tsx
@@ -1,10 +1,10 @@
-import { EditProfileForm, SubscriptionPlans } from '../../molecules';
+import { DesktopEditProfileForm, SubscriptionPlans } from '../../molecules';
 import { IProfileProps } from '../../profile.types';
 import S from './mobile-profile.styles';
 
 export const MobileProfile = ({ plans, ...restProps }: IProfileProps): JSX.Element => (
   <S.StyledWrapper>
     <SubscriptionPlans plans={plans} />
-    <EditProfileForm {...restProps} />
+    <DesktopEditProfileForm {...restProps} />
   </S.StyledWrapper>
 );
